fix(gr00t_n1d7): refactor gr00t_n1d7; fix bug of install.sh
