+++ rlinf/envs/action_utils.py
@@ -72,6 +72,7 @@ def prepare_actions_for_libero(
         SupportedModel.OPENVLA,
         SupportedModel.OPENVLA_OFT,
         SupportedModel.GR00T_N1D6,
+        SupportedModel.GR00T_N1D7,
     ]:
         chunk_actions[..., -1] = 2 * chunk_actions[..., -1] - 1
         chunk_actions[..., -1] = np.sign(chunk_actions[..., -1]) * -1.0
