+++ rlinf/models/embodiment/gr00t/simulation_io.py
@@ -1,4 +1,4 @@
-# Copyright 2025 The RLinf Authors.
+# Copyright 2026 The RLinf Authors.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -48,6 +48,7 @@ def convert_maniskill_obs_to_gr00t_format(env_obs):
     following LeRobot format.
     """
     groot_obs = {}
+    # video
     # TODO(lx): If we have a dataset on maniskill, resize can be avoided.
     # But now we have to resize images to libero data version.
     env_obs["main_images"] = cut_and_resize_images(
@@ -57,10 +58,14 @@ def convert_maniskill_obs_to_gr00t_format(env_obs):
     )
     # [B, H, W, C] -> [B, T, H, W, C]
     groot_obs["video.ego_view"] = env_obs["main_images"].unsqueeze(1).numpy()
+    # state
     if "state" in env_obs:
         raise NotImplementedError("State from simulation are not unified yet.")
-    # Pad state to input dimension: [B, T, C]
-    groot_obs["state.left_arm"] = np.zeros((env_obs["main_images"].shape[0], 1, 7))
+    else:
+        # gr00t_1_7 pad the state to input dimension
+        # create state of [B, T, C]
+        groot_obs["state.left_arm"] = np.zeros((env_obs["main_images"].shape[0], 1, 7))
+    # annotation
     groot_obs["annotation.human.action.task_description"] = env_obs["task_descriptions"]
     return groot_obs
 
@@ -130,6 +135,51 @@ def convert_to_libero_action_n1d6(
     return action_array
 
 
+def convert_to_libero_action_n1d7(
+    action_chunk: dict[str, np.array],
+    chunk_size: int = 1,
+) -> np.ndarray:
+    """Convert GR00T N1.7 action chunk to a 7-dim Libero action array.
+
+    Gripper normalization is NOT applied here; it is handled by the shared
+    ``prepare_actions_for_libero`` in ``rlinf.envs.action_utils``.
+    """
+    try:
+        pos = action_chunk["end_effector_position"][:, :chunk_size]
+        rot = action_chunk["end_effector_rotation"][:, :chunk_size]
+        gripper = action_chunk["gripper_close"][:, :chunk_size]
+        action_array = np.concatenate([pos, rot, gripper], axis=-1)
+    except KeyError:
+        if all(
+            key in action_chunk
+            for key in ("x", "y", "z", "roll", "pitch", "yaw", "gripper")
+        ):
+            action_array = np.concatenate(
+                [
+                    action_chunk["x"][:, :chunk_size],
+                    action_chunk["y"][:, :chunk_size],
+                    action_chunk["z"][:, :chunk_size],
+                    action_chunk["roll"][:, :chunk_size],
+                    action_chunk["pitch"][:, :chunk_size],
+                    action_chunk["yaw"][:, :chunk_size],
+                    action_chunk["gripper"][:, :chunk_size],
+                ],
+                axis=-1,
+            )
+        elif "rel_arm_action" in action_chunk:
+            arm = action_chunk["rel_arm_action"][:, :chunk_size]
+            grp = action_chunk["gripper_action"][:, :chunk_size]
+            action_array = np.concatenate([arm, grp], axis=-1)
+        else:
+            raise KeyError(f"can not find Action Keys: {list(action_chunk.keys())}")
+    # gripper conversion is handled by the shared
+    # ``prepare_actions_for_libero`` in ``rlinf.envs.action_utils``.
+    assert action_array.shape[-1] == 7, (
+        f"Expected 7-dim action, got {action_array.shape[-1]}"
+    )
+    return action_array
+
+
 def convert_to_maniskill_action(
     action_chunk: dict[str, np.array], chunk_size: int = 16
 ) -> np.ndarray:
@@ -176,6 +226,12 @@ ACTION_CONVERSION_N1D6 = {
     "isaaclab_stack_cube": convert_to_isaaclab_stack_cube_action,
 }
 
+ACTION_CONVERSION_N1D7 = {
+    "libero": convert_to_libero_action_n1d7,
+    "maniskill": convert_to_maniskill_action,
+    "isaaclab_stack_cube": convert_to_isaaclab_stack_cube_action,
+}
+
 
 def cut_and_resize_images(
     images: torch.Tensor, crop_size: int, target_size: int = 256
@@ -183,20 +239,25 @@ def cut_and_resize_images(
     """Cut and resize the images to the crop size."""
     images_nchw = images.permute(0, 3, 1, 2)  # [B, H, W, C] -> [B, C, H, W]
 
-    original_width = images_nchw.shape[-1]
+    original_width = images_nchw.shape[-1]  # W
     start = (original_width - crop_size) // 2
     end = start + crop_size
 
-    cropped_tensor = images_nchw[:, :, :, start:end]
+    # Crop: keep batch, channels, full height; crop width to [start:end]
+    cropped_tensor = images_nchw[:, :, :, start:end]  # [B, C, H, crop_W]
 
+    # Resize: interpolate to target_size x target_size
     resized_tensor = F.interpolate(
         cropped_tensor,
         size=(target_size, target_size),
-        mode="bilinear",
+        mode="bilinear",  # Or 'bicubic' for smoother results
         align_corners=False,
-    )
+    )  # [B, C, target_size, target_size]
 
-    resized_nhwc = resized_tensor.permute(0, 2, 3, 1).contiguous()
+    # Convert back to NHWC
+    resized_nhwc = resized_tensor.permute(
+        0, 2, 3, 1
+    ).contiguous()  # [B, C, H, W] -> [B, H, W, C]
     return resized_nhwc
