--- rlinf/models/embodiment/gr00t_1_7/embodiment_tags.py
@@ -1,99 +1,0 @@
-# Copyright 2026 The RLinf Authors.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-
-from enum import Enum
-
-
-class EmbodimentTag(Enum):
-    AGIBOT_GENIE1 = "agibot_genie1"
-    """
-    The AgiBot Genie-1 with gripper dataset.
-    """
-
-    LIBERO_FRANKA = "libero_franka"
-    """
-    The Libero Franka dataset.
-    """
-
-    MANISKILL_WIDOWX = "maniskill_widowx"
-    """
-    The maniskill widowx dataset.
-    """
-
-    ISAACLAB_FRANKA = "isaaclab_franka"
-    """
-    The isaaclab Franka dataset.
-    """
-    # Pretrain embodiment tags #####
-    ROBOCASA_PANDA_OMRON = "robocasa_panda_omron"
-    """
-    The RoboCasa Panda robot with omron mobile base.
-    """
-
-    GR1 = "gr1"
-    """
-    The Fourier GR1 robot.
-    """
-
-    # Pre-registered posttrain embodiment tags #####
-    UNITREE_G1 = "unitree_g1"
-    """
-    The Unitree G1 robot.
-    """
-
-    LIBERO_PANDA = "libero_sim"
-    """
-    The Libero panda robot.
-    """
-
-    OXE_GOOGLE = "oxe_google"
-    """
-    The Open-X-Embodiment Google robot.
-    """
-
-    OXE_WIDOWX = "oxe_widowx"
-    """
-    The Open-X-Embodiment WidowX robot.
-    """
-
-    OXE_DROID = "oxe_droid"
-    """
-    The Open-X-Embodiment DROID robot with relative joint position actions.
-    """
-
-    BEHAVIOR_R1_PRO = "behavior_r1_pro"
-    """
-    The Behavior R1 Pro robot.
-    """
-
-    # New embodiment during post-training
-    NEW_EMBODIMENT = "new_embodiment"
-    """
-    Any new embodiment.
-    """
-
-
-# Embodiment tag string to projector index in the Action Expert Module.
-# These IDs must match the official GR00T processor/checkpoint metadata.
-EMBODIMENT_TAG_MAPPING = {
-    EmbodimentTag.LIBERO_PANDA.value: 2,
-    EmbodimentTag.ROBOCASA_PANDA_OMRON.value: 13,
-    EmbodimentTag.LIBERO_FRANKA.value: 31,
-    EmbodimentTag.OXE_DROID.value: 17,
-    EmbodimentTag.AGIBOT_GENIE1.value: 26,
-    EmbodimentTag.GR1.value: 24,
-    EmbodimentTag.MANISKILL_WIDOWX.value: 30,
-    EmbodimentTag.ISAACLAB_FRANKA.value: 31,
-}
