--- rlinf/models/embodiment/gr00t_1_7/simulation_io.py
@@ -1,264 +1,0 @@
-# Copyright 2026 The RLinf Authors.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import os
-
-import numpy as np
-import torch
-import torch.nn.functional as F
-
-
-def convert_libero_obs_to_gr00t_1_7_format(env_obs):
-    """
-    Convert the observation to the format expected by the gr00t_1_7 model.
-    The data format is determined by the modality_config and meta/info.json following LeRobot format.
-    Considering that we don't have a unified data inferface, we use direct logic here.
-    """
-    groot_obs = {}
-
-    # [B, H, W, C] -> [B, T, H, W, C]
-    groot_obs["video.image"] = env_obs["main_images"].unsqueeze(1).numpy()
-    groot_obs["video.wrist_image"] = env_obs["wrist_images"].unsqueeze(1).numpy()
-    # [B, 8] -> [B, T(1), 8]
-    groot_obs["state.x"] = env_obs["states"].unsqueeze(1)[:, :, 0:1].numpy()
-    groot_obs["state.y"] = env_obs["states"].unsqueeze(1)[:, :, 1:2].numpy()
-    groot_obs["state.z"] = env_obs["states"].unsqueeze(1)[:, :, 2:3].numpy()
-    groot_obs["state.roll"] = env_obs["states"].unsqueeze(1)[:, :, 3:4].numpy()
-    groot_obs["state.pitch"] = env_obs["states"].unsqueeze(1)[:, :, 4:5].numpy()
-    groot_obs["state.yaw"] = env_obs["states"].unsqueeze(1)[:, :, 5:6].numpy()
-    groot_obs["state.gripper"] = env_obs["states"].unsqueeze(1)[:, :, 6:].numpy()
-    groot_obs["annotation.human.action.task_description"] = env_obs["task_descriptions"]
-
-    return groot_obs
-
-
-def convert_maniskill_obs_to_gr00t_1_7_format(env_obs):
-    """
-    Convert the observation to the format expected by the gr00t_1_7 model.
-    The data format is determined by the modality_config and meta/info.json following LeRobot format.
-    Considering that we don't have a unified data inferface, we use direct logic here.
-    """
-    groot_obs = {}
-    # video
-    # TODO(lx): If we have a dataset on maniskill, resize can be avoided.
-    # But now we have to resize images to libero data version.
-    env_obs["main_images"] = cut_and_resize_images(
-        env_obs["main_images"],
-        env_obs["main_images"].shape[-3],  # H
-        256,
-    )
-    # [B, H, W, C] -> [B, T, H, W, C]
-    groot_obs["video.ego_view"] = env_obs["main_images"].unsqueeze(1).numpy()
-    # state
-    if "state" in env_obs:
-        raise NotImplementedError("State from simulation are not unified yet.")
-    else:
-        # gr00t_1_7 pad the state to input dimension
-        # create state of [B, T, C]
-        groot_obs["state.left_arm"] = np.zeros((env_obs["main_images"].shape[0], 1, 7))
-    # annotation
-    groot_obs["annotation.human.action.task_description"] = env_obs["task_descriptions"]
-    return groot_obs
-
-
-def _debug_libero_action_stats(prefix: str, action_array: np.ndarray):
-    dim_names = ["x", "y", "z", "roll", "pitch", "yaw", "gripper"]
-    flat = action_array.reshape(-1, action_array.shape[-1]).astype(np.float32)
-    stats = []
-    for dim, name in enumerate(dim_names[: flat.shape[-1]]):
-        vals = flat[:, dim]
-        stats.append(
-            f"{name}:min={vals.min():.6g},max={vals.max():.6g},mean={vals.mean():.6g}"
-        )
-    preview = action_array[0, : min(action_array.shape[1], 8)].astype(np.float32)
-    print(
-        f"[GR00T_LIBERO_ACTION_DEBUG] {prefix}: shape={action_array.shape} "
-        f"{' | '.join(stats)} first_chunk={preview.tolist()}",
-        flush=True,
-    )
-
-
-def _apply_official_libero_gripper_action(action_array: np.ndarray) -> np.ndarray:
-    # Match Isaac-GR00T official LIBERO wrapper: normalize [0, 1] -> [-1, 1],
-    # binarize, then invert because LIBERO uses -1=open and +1=close.
-    action_array = action_array.copy()
-    action_array[..., -1] = 2 * action_array[..., -1] - 1
-    action_array[..., -1] = np.sign(action_array[..., -1])
-    action_array[..., -1] *= -1.0
-    return action_array
-
-
-def convert_to_libero_action(
-    action_chunk: dict[str, np.array],
-    chunk_size: int = 1,
-    gripper_mode: str = "legacy",
-) -> np.ndarray:
-    try:
-        pos = action_chunk["end_effector_position"][:, :chunk_size]  # (B, chunk, 3)
-        rot = action_chunk["end_effector_rotation"][:, :chunk_size]  # (B, chunk, 3)
-        gripper = action_chunk["gripper_close"][:, :chunk_size]  # (B, chunk, 1)
-
-        action_array = np.concatenate([pos, rot, gripper], axis=-1)
-
-    except KeyError:
-        if all(
-            key in action_chunk
-            for key in ("x", "y", "z", "roll", "pitch", "yaw", "gripper")
-        ):
-            action_array = np.concatenate(
-                [
-                    action_chunk["x"][:, :chunk_size],
-                    action_chunk["y"][:, :chunk_size],
-                    action_chunk["z"][:, :chunk_size],
-                    action_chunk["roll"][:, :chunk_size],
-                    action_chunk["pitch"][:, :chunk_size],
-                    action_chunk["yaw"][:, :chunk_size],
-                    action_chunk["gripper"][:, :chunk_size],
-                ],
-                axis=-1,
-            )
-        elif "rel_arm_action" in action_chunk:
-            arm = action_chunk["rel_arm_action"][:, :chunk_size]
-            grp = action_chunk["gripper_action"][:, :chunk_size]
-            action_array = np.concatenate([arm, grp], axis=-1)
-        else:
-            raise KeyError(f"can not find Action Keys: {list(action_chunk.keys())}")
-
-    if os.getenv("RLINF_DEBUG_GR00T_ACTION") == "1" and not getattr(
-        convert_to_libero_action, "_printed_pre_debug", False
-    ):
-        convert_to_libero_action._printed_pre_debug = True
-        _debug_libero_action_stats("pre_gripper_conversion", action_array)
-
-    if gripper_mode == "legacy":
-        action_array = normalize_gripper_action(action_array, binarize=True)
-    elif gripper_mode == "official":
-        action_array = _apply_official_libero_gripper_action(action_array)
-    elif gripper_mode == "passthrough":
-        action_array = action_array.copy()
-        action_array[..., -1] = np.sign(action_array[..., -1])
-    else:
-        raise ValueError(
-            f"Unknown libero gripper conversion mode: {gripper_mode}. "
-            "Expected one of: legacy, official, passthrough."
-        )
-
-    if os.getenv("RLINF_DEBUG_GR00T_ACTION") == "1" and not getattr(
-        convert_to_libero_action, "_printed_post_debug", False
-    ):
-        convert_to_libero_action._printed_post_debug = True
-        _debug_libero_action_stats(
-            f"post_gripper_conversion[{gripper_mode}]", action_array
-        )
-
-    return action_array
-
-
-def convert_to_maniskill_action(
-    action_chunk: dict[str, np.array], chunk_size: int = 16
-) -> np.ndarray:
-    """Convert gr00t_1_7 action chunk to Maniskill format."""
-    # Accord to gr1 definition, action.left_arm happens to be 7 dims, matching the demand of maniskill.
-
-    return action_chunk["action.left_arm"][:, :chunk_size]
-
-
-def convert_to_isaaclab_stack_cube_action(
-    action_chunk: dict[str, np.array], chunk_size: int = 1
-) -> np.ndarray:
-    """Convert gr00t_1_7 action chunk to Isaaclab Stack Cube format.
-    The main difference of Libero and Isaaclab Stack Cube is gripper action in
-    Libero is 0 and 1, but in Isaaclab Stack Cube is -1 and +1.
-
-    Args:
-        action_chunk: Dictionary of action components from gr00t_1_7 policy
-        chunk_size: Number of action steps to consider from the chunk
-
-    Returns:
-        7-dim numpy array: [dx, dy, dz, droll, dpitch, dyaw, gripper]
-    """
-    action_components = [
-        action_chunk["action.x"][:, :chunk_size],
-        action_chunk["action.y"][:, :chunk_size],
-        action_chunk["action.z"][:, :chunk_size],
-        action_chunk["action.roll"][:, :chunk_size],
-        action_chunk["action.pitch"][:, :chunk_size],
-        action_chunk["action.yaw"][:, :chunk_size],
-        action_chunk["action.gripper"][:, :chunk_size],
-    ]
-    action_array = np.concatenate(action_components, axis=-1)
-    action_array[..., -1] = np.sign(action_array[..., -1])
-    assert action_array.shape[-1] == 7, (
-        f"Expected 7-dim action, got {action_array.shape[-1]}"
-    )
-    return action_array
-
-
-# TODO: we need a unified embodiement data.
-OBS_CONVERSION = {
-    "maniskill": convert_maniskill_obs_to_gr00t_1_7_format,
-    "libero": convert_libero_obs_to_gr00t_1_7_format,
-    "isaaclab_stack_cube": convert_libero_obs_to_gr00t_1_7_format,
-}
-
-ACTION_CONVERSION = {
-    "libero": convert_to_libero_action,
-    "maniskill": convert_to_maniskill_action,
-    "isaaclab_stack_cube": convert_to_isaaclab_stack_cube_action,
-}
-
-
-def cut_and_resize_images(
-    images: torch.Tensor, crop_size: int, target_size: int = 256
-) -> torch.Tensor:
-    """
-    Cut and resize the images to the crop size.
-    """
-    images_nchw = images.permute(0, 3, 1, 2)  # [B, H, W, C] -> [B, C, H, W]
-
-    original_width = images_nchw.shape[-1]  # W
-    start = (original_width - crop_size) // 2
-    end = start + crop_size
-
-    # Crop: keep batch, channels, full height; crop width to [start:end]
-    cropped_tensor = images_nchw[:, :, :, start:end]  # [B, C, H, crop_W]
-
-    # Resize: interpolate to target_size x target_size
-    resized_tensor = F.interpolate(
-        cropped_tensor,
-        size=(target_size, target_size),
-        mode="bilinear",  # Or 'bicubic' for smoother results
-        align_corners=False,
-    )  # [B, C, target_size, target_size]
-
-    # Convert back to NHWC
-    resized_nhwc = resized_tensor.permute(
-        0, 2, 3, 1
-    ).contiguous()  # [B, C, H, W] -> [B, H, W, C]
-    return resized_nhwc
-
-
-def normalize_gripper_action(action, binarize=True):
-    """
-    Changes gripper action (last dimension of action vector) from [0,1] to [+1,-1].
-
-    Normalization formula: y = 1 - 2 * (x - orig_low) / (orig_high - orig_low)
-    """
-    orig_low, orig_high = 0.0, 1.0
-    action[..., -1] = 1 - 2 * (action[..., -1] - orig_low) / (orig_high - orig_low)
-
-    if binarize:
-        action[..., -1] = np.sign(action[..., -1])
-
-    return action
