+++ tests/e2e_tests/embodied/libero_spatial_ppo_gr00t_n1d7.yaml
@@ -16,12 +16,12 @@ hydra:
 cluster:
   num_nodes: 1
   component_placement:
-    actor,env,rollout: all
+    actor,env,rollout: 0-1
 
 runner:
   task_type: embodied
   logger:
-    log_path: "../results"
+    log_path: "../results/libero_spatial_ppo_gr00t_n1d7"
     project_name: rlinf
     experiment_name: "test_gr00t_n1d7"
     logger_backends: ["tensorboard"] # wandb, swanlab
@@ -49,7 +49,7 @@ algorithm:
   adv_type: gae
   loss_type: actor_critic
   loss_agg_func: "token-mean" 
-  kl_beta: 0.0
+  kl_beta: 0.01
   entropy_bonus: 0
   clip_ratio_high: 0.2
   clip_ratio_low: 0.2
@@ -88,6 +88,7 @@ env:
     total_num_envs: 8
     max_episode_steps: 5
     max_steps_per_rollout_epoch: 5
+    auto_reset: True
   eval:
     total_num_envs: 1
     max_episode_steps: 10
@@ -106,8 +107,9 @@ rollout:
   backend: "huggingface"
   enable_offload: True
   pipeline_stage_num: 1
-  model:  
+  model:
     model_path: /workspace/dataset/GR00T-N1.7-LIBERO/libero_spatial
+    backbone_model_path: /workspace/dataset/Cosmos-Reason2-2B
     precision: ${actor.model.precision}
 
 actor:
@@ -121,9 +123,12 @@ actor:
   # Override the default values in model/gr00t_n1d7
   model:
     model_path: /workspace/dataset/GR00T-N1.7-LIBERO/libero_spatial
+    backbone_model_path: /workspace/dataset/Cosmos-Reason2-2B
     model_type: "gr00t_n1d7"
+    embodiment_tag: "libero_sim"
     add_value_head: True
-    num_action_chunks: 1       
+    num_action_chunks: 1
+    action_dim: 7
     denoising_steps: 1
     no_split_modules:
       - "Qwen3DecoderLayer"
@@ -132,7 +137,7 @@ actor:
       - "TimestepEncoder"
       - "TimestepEmbedding"
       - "ValueHead"
-      - "ExploreNoiseNet"        
+      - "ExploreNoiseNet"
 
   optim:
     lr: 5.0e-6
